Moved manager and handler to the common package.  Removed the abstract handler as it isn't needed.  Renames to EntityParentIdServiceExtension...
