--- admin/broadleaf-open-admin-platform/src/main/java/org/broadleafcommerce/openadmin/handler/AbstractParentIdServiceExtensionHandler.java
@@ -1,31 +1,0 @@
-/*
- * #%L
- * BroadleafCommerce Framework
- * %%
- * Copyright (C) 2009 - 2016 Broadleaf Commerce
- * %%
- * Licensed under the Broadleaf Fair Use License Agreement, Version 1.0
- * (the "Fair Use License" located  at http://license.broadleafcommerce.org/fair_use_license-1.0.txt)
- * unless the restrictions on use therein are violated and require payment to Broadleaf in which case
- * the Broadleaf End User License Agreement (EULA), Version 1.1
- * (the "Commercial License" located at http://license.broadleafcommerce.org/commercial_license-1.1.txt)
- * shall apply.
- * 
- * Alternatively, the Commercial License may be replaced with a mutually agreed upon license (the "Custom License")
- * between you and Broadleaf Commerce. You may not use this file except in compliance with the applicable license.
- * #L%
- */
-package org.broadleafcommerce.openadmin.handler;
-
-import org.broadleafcommerce.common.extension.AbstractExtensionHandler;
-import org.broadleafcommerce.common.extension.ExtensionResultHolder;
-import org.broadleafcommerce.common.extension.ExtensionResultStatusType;
-
-public class AbstractParentIdServiceExtensionHandler extends AbstractExtensionHandler implements BasicParentIdServiceExtensionHandler {
-
-    @Override
-    public ExtensionResultStatusType checkParentId(String id, String className, ExtensionResultHolder<String> resultHolder) {
-        return ExtensionResultStatusType.NOT_HANDLED;
-    }
-}
-
